main/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
distro==1.9.0
Django==5.1.7
django-cors-headers==4.7.0
django-environ==0.12.0
django-mongodb-backend==5.1.0b1
djangorestframework==3.15.2
dnspython==2.7.0
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
openai==1.69.0
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.0
pydantic_core==2.33.0
pymongo==4.11.3
requests==2.32.3
rsa==4.9
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
whitenoise==6.9.0
